+++ R/generic-s3.R
@@ -11,8 +11,7 @@
 #' Plots an \code{aggTrees} object.
 #'
 #' @examples
-#' \donttest{
-#' ## Generate data.
+#' \donttest{## Generate data.
 #' set.seed(1986)
 #'
 #' n <- 1000
@@ -23,10 +22,25 @@
 #' D <- rbinom(n, size = 1, prob = 0.5)
 #' mu0 <- 0.5 * X[, 1]
 #' mu1 <- 0.5 * X[, 1] + X[, 2]
-#' y <- mu0 + D * (mu1 - mu0) + rnorm(n)
+#' Y <- mu0 + D * (mu1 - mu0) + rnorm(n)
 #'
-#' ## Construct sequence of groupings. CATEs estimated internally,
-#' groupings <- build_aggtree(y, D, X, method = "aipw")
+#' ## Training-honest sample split.
+#' honest_frac <- 0.5
+#' splits <- sample_split(length(Y), training_frac = (1 - honest_frac))
+#' training_idx <- splits$training_idx
+#' honest_idx <- splits$honest_idx
+#'
+#' Y_tr <- Y[training_idx]
+#' D_tr <- D[training_idx]
+#' X_tr <- X[training_idx, ]
+#'
+#' Y_hon <- Y[honest_idx]
+#' D_hon <- D[honest_idx]
+#' X_hon <- X[honest_idx, ]
+#'
+#' ## Construct sequence of groupings. CATEs estimated internally.
+#' groupings <- build_aggtree(Y_tr, D_tr, X_tr,
+#'                            Y_hon, D_hon, X_hon)
 #'
 #' ## Plot.
 #' plot(groupings)
@@ -44,7 +58,7 @@
 #'
 #' @references
 #' \itemize{
-#'   \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
+#'   \item Di Francesco, R. (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
 #' }
 #'
 #' @seealso
@@ -176,14 +190,14 @@ plot.aggTrees <- function(x, leaves = get_leaves(x$tree), sequence = FALSE, ...)
 #'
 #' @references
 #' \itemize{
-#'   \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
+#'   \item Di Francesco, R. (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
 #' }
 #'
 #' @author Riccardo Di Francesco
 #'
 #' @export
 summary.aggTrees <- function(object, ...) {
-  if (is.null(object$idx$honest_idx)) cat("Honest estimates:", FALSE, "\n") else cat("Honest estimates:", TRUE, "\n")
+  if (is.null(object$honest_sample)) cat("Honest estimates:", FALSE, "\n") else cat("Honest estimates:", TRUE, "\n")
   summary(object$tree)
 }
 
@@ -203,14 +217,14 @@ summary.aggTrees <- function(object, ...) {
 #'
 #' @references
 #' \itemize{
-#'   \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
+#'   \item Di Francesco, R. (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
 #' }
 #'
 #' @author Riccardo Di Francesco
 #'
 #' @export
 print.aggTrees <- function(x, ...) {
-  if (is.null(x$idx$honest_idx)) cat("Honest estimates:", FALSE, "\n") else cat("Honest estimates:", TRUE, "\n")
+  if (is.null(x$honest_sample)) cat("Honest estimates:", FALSE, "\n") else cat("Honest estimates:", TRUE, "\n")
   print(x$tree)
 }
 
@@ -227,8 +241,7 @@ print.aggTrees <- function(x, ...) {
 #' Prints LATEX code.
 #'
 #' @examples
-#' \donttest{
-#' ## Generate data.
+#' \donttest{## Generate data.
 #' set.seed(1986)
 #'
 #' n <- 1000
@@ -239,10 +252,25 @@ print.aggTrees <- function(x, ...) {
 #' D <- rbinom(n, size = 1, prob = 0.5)
 #' mu0 <- 0.5 * X[, 1]
 #' mu1 <- 0.5 * X[, 1] + X[, 2]
-#' y <- mu0 + D * (mu1 - mu0) + rnorm(n)
+#' Y <- mu0 + D * (mu1 - mu0) + rnorm(n)
 #'
-#' ## Construct sequence of groupings. CATEs estimated internally,
-#' groupings <- build_aggtree(y, D, X, method = "aipw")
+#' ## Training-honest sample split.
+#' honest_frac <- 0.5
+#' splits <- sample_split(length(Y), training_frac = (1 - honest_frac))
+#' training_idx <- splits$training_idx
+#' honest_idx <- splits$honest_idx
+#'
+#' Y_tr <- Y[training_idx]
+#' D_tr <- D[training_idx]
+#' X_tr <- X[training_idx, ]
+#'
+#' Y_hon <- Y[honest_idx]
+#' D_hon <- D[honest_idx]
+#' X_hon <- X[honest_idx, ]
+#'
+#' ## Construct sequence of groupings. CATEs estimated internally.
+#' groupings <- build_aggtree(Y_tr, D_tr, X_tr,
+#'                            Y_hon, D_hon, X_hon)
 #'
 #' ## Analyze results with 4 groups.
 #' results <- inference_aggtree(groupings, n_groups = 4)
@@ -266,7 +294,7 @@ print.aggTrees <- function(x, ...) {
 #'
 #' @references
 #' \itemize{
-#'   \item R Di Francesco (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
+#'   \item Di Francesco, R. (2022). Aggregation Trees. CEIS Research Paper, 546. \doi{10.2139/ssrn.4304256}.
 #' }
 #'
 #' @author Riccardo Di Francesco
@@ -277,10 +305,10 @@ print.aggTrees.inference <- function(x, table = "avg_char", ...) {
   if (!(table %in% c("avg_char", "diff"))) stop("Invalid 'table'. This must be either 'avg_char' or 'diff'.", call. = FALSE)
 
   ## Select appropriate sample (adaptive/honest) according to the output of build_aggtree.
-  if (is.null(x$idx$honest_idx)) {
-    X <- x$aggTree$dta[, -c(1,2 )]
+  if (is.null(x$aggTree$honest_sample)) {
+    X <- x$aggTree$training_sample[, -c(1:3)]
   } else {
-    X <- x$aggTree$dta[x$aggTree$idx$honest_idx, -c(1,2 )]
+    X <- x$aggTree$honest_sample[, -c(1:3)]
   }
 
   ## Extract information.
@@ -293,14 +321,14 @@ print.aggTrees.inference <- function(x, table = "avg_char", ...) {
     gates_idx <- which(sapply(names(x$model$coefficients), function(x) grepl("leaf", x)))
   }
 
-  gates_point <- round(x$model$coefficients[gates_idx], 3)
-  gates_sd <- round(x$model$std.error[gates_idx], 3)
+  gates_point <- format(round(x$model$coefficients[gates_idx], 3), nsmall = 3)
+  gates_sd <- format(round(x$model$std.error[gates_idx], 3), nsmall = 3)
 
-  gates_ci_lower <- round(gates_point - 1.96 * gates_sd, 3)
-  gates_ci_upper <- round(gates_point + 1.96 * gates_sd, 3)
+  gates_ci_lower <- format(round(as.numeric(gates_point) - 1.96 * as.numeric(gates_sd), 3), nsmall = 3)
+  gates_ci_upper <- format(round(as.numeric(gates_point) + 1.96 * as.numeric(gates_sd), 3), nsmall = 3)
 
-  if (length(x$boot_ci) != 0) gates_ci_lower_boot <- round(x$boot_ci$lower, 3) else gates_ci_lower_boot <- NA
-  if (length(x$boot_ci) != 0) gates_ci_upper_boot <- round(x$boot_ci$upper, 3) else gates_ci_upper_boot <- NA
+  if (length(x$boot_ci) != 0) gates_ci_lower_boot <- format(round(x$boot_ci$lower, 3), nsmall = 3) else gates_ci_lower_boot <- NA
+  if (length(x$boot_ci) != 0) gates_ci_upper_boot <- format(round(x$boot_ci$upper, 3), nsmall = 3) else gates_ci_upper_boot <- NA
 
   ## Write table.
   if (table == "avg_char") {
@@ -321,8 +349,8 @@ print.aggTrees.inference <- function(x, table = "avg_char", ...) {
       \\hline \\\\[-1.8ex] \n\n", sep = "")
 
     for (i in seq_len(length(table_names))) {
-      temp_means <- round(parms[[i]][, 1], 3)
-      temp_sds <- round(parms[[i]][, 2], 3)
+      temp_means <- format(round(parms[[i]][, 1], 3), nsmall = 3)
+      temp_sds <- format(round(parms[[i]][, 2], 3), nsmall = 3)
 
       temp_line <- paste0("      \\texttt{", table_names[i], "} & " )
 
@@ -340,8 +368,8 @@ print.aggTrees.inference <- function(x, table = "avg_char", ...) {
       \\hline \\\\[-1.8ex]
     \\end{tabular}
     \\end{adjustbox}
-    \\caption{Average characteristics of units in each leaf, obtained by regressing each covariate on a set of dummies denoting leaf membership. Standard errors are estimated via the Eicker-Huber-White estimator. Leaves are sorted in increasing order of the GATEs.}
-    \\label{table:average.characteristics.leaves}
+    \\caption{Average characteristics of units in each leaf, obtained by regressing each covariate on a set of dummies denoting leaf membership", if (!is.null(x$aggTree$idx$honest_idx)) "using only the honest sample." else "." ,"Standard errors are estimated via the Eicker-Huber-White estimator. Leaves are sorted in increasing order of the GATEs.}
+    \\label{table_average_characteristics_leaves}
     \\end{table}
 \\endgroup \n\n")
 
@@ -360,7 +388,7 @@ print.aggTrees.inference <- function(x, table = "avg_char", ...) {
   \\renewcommand{\\arraystretch}{1.2}
   \\begin{table}[b!]
     \\centering
-    \\begin{adjustbox}{width = 0.85\\textwidth}
+    \\begin{adjustbox}{width = 1\\textwidth}
     \\begin{tabular}{@{\\extracolsep{5pt}}l", rep(" c", times = get_leaves(x$groups)), "}
       \\\\[-1.8ex]\\hline
       \\hline \\\\[-1.8ex] \n
@@ -370,13 +398,13 @@ print.aggTrees.inference <- function(x, table = "avg_char", ...) {
 
     cat("      \\multirow{3}{*}{GATEs} & ", paste(gates_point[1:(length(unique(leaves))-1)], " & ", sep = ""), gates_point[length(unique(leaves))], " \\\\
       & ", paste("[", gates_ci_lower[1:(length(unique(leaves))-1)], ", ", gates_ci_upper[1:(length(unique(leaves))-1)], "] & ", sep = ""), paste("[", gates_ci_lower[length(unique(leaves))], ", ", gates_ci_upper[length(unique(leaves))], "]", sep = ""), " \\\\
-      & ", paste("[", gates_ci_lower_boot[1:(length(unique(leaves))-1)], ", ", gates_ci_upper_boot[1:(length(unique(leaves))-1)], "] & ", sep = ""), paste("[", gates_ci_lower_boot[length(unique(leaves))], ", ", gates_ci_upper_boot[length(unique(leaves))], "]", sep = ""), " \\\\ \n\n", sep = "")
+      & ", paste("\\{", gates_ci_lower_boot[1:(length(unique(leaves))-1)], ", ", gates_ci_upper_boot[1:(length(unique(leaves))-1)], "\\} & ", sep = ""), paste("\\{", gates_ci_lower_boot[length(unique(leaves))], ", ", gates_ci_upper_boot[length(unique(leaves))], "\\}", sep = ""), " \\\\ \n\n", sep = "")
     cat("      \\addlinespace[2pt]
       \\hline \\\\[-1.8ex] \n\n")
 
     for (i in seq_len(get_leaves(x$groups))) {
-      cat(paste0("      \\textit{Leaf ", i, "}"), " & ", paste0(round(x$gates_diff_pairs$gates_diff[i, seq_len(get_leaves(x$groups)-1)], 3), " & "), round(x$gates_diff_pairs$gates_diff[i, get_leaves(x$groups)], 3), " \\\\
-            & ", paste0("(", round(x$gates_diff_pairs$holm_pvalues[i, seq_len(get_leaves(x$groups)-1)], 3), ") & "), paste0("(", round(x$gates_diff_pairs$holm_pvalues[i, get_leaves(x$groups)], 3), ")"), " \\\\ \n", sep = "")
+      cat(paste0("      \\textit{Leaf ", i, "}"), " & ", paste0(format(round(x$gates_diff_pairs$gates_diff[i, seq_len(get_leaves(x$groups)-1)], 2), nsmall = 2), " & "), format(round(x$gates_diff_pairs$gates_diff[i, get_leaves(x$groups)], 2), nsmall = 2), " \\\\
+            & ", paste0("(", format(round(x$gates_diff_pairs$holm_pvalues[i, seq_len(get_leaves(x$groups)-1)], 3), nsmall = 3), ") & "), paste0("(", format(round(x$gates_diff_pairs$holm_pvalues[i, get_leaves(x$groups)], 3), nsmall = 3), ")"), " \\\\ \n", sep = "")
     }
 
     cat("\n      \\addlinespace[3pt]
@@ -384,8 +412,8 @@ print.aggTrees.inference <- function(x, table = "avg_char", ...) {
       \\hline \\\\[-1.8ex]
     \\end{tabular}
     \\end{adjustbox}
-    \\caption{Point estimates and $95\\%$ confidence intervals for the GATEs based on asymptotic normality and on the percentiles of the bootstrap distribution. Leaves are sorted in increasing order of the GATEs. Additionally, differences in the GATEs across all pairs of leaves are displayed. p-values to test the null hypothesis that a single difference is zero are adjusted using Holm's procedure and reported in parenthesis under each point estimate.}
-    \\label{table:differences.gates}
+    \\caption{Point estimates and $95\\%$ confidence intervals for the GATEs based on asymptotic normality (in square brackets) and on the percentiles of the bootstrap distribution (in curly braces). Leaves are sorted in increasing order of the GATEs. Additionally, the GATE differences across all pairs of leaves are displayed. $p$-values testing the null hypothesis that a single difference is zero are adjusted using Holm's procedure and reported in parenthesis under each point estimate.}
+    \\label{table_differences_gates}
     \\end{table}
 \\endgroup \n\n")
   }