+++ inst/doc/aggTrees-vignette.R
@@ -5,56 +5,60 @@ knitr::opts_chunk$set(
 )
 
 library(aggTrees)
+library(grf)
 
-## ----data-generation, eval = FALSE--------------------------------------------
-#  ## Generate data.
-#  set.seed(1986)
-#  
-#  n <- 1000
-#  k <- 3
-#  
-#  X <- matrix(rnorm(n * k), ncol = k)
-#  colnames(X) <- paste0("x", seq_len(k))
-#  D <- rbinom(n, size = 1, prob = 0.5)
-#  mu0 <- 0.5 * X[, 1]
-#  mu1 <- 0.5 * X[, 1] + X[, 2]
-#  y <- mu0 + D * (mu1 - mu0) + rnorm(n)
+## ----data-generation, eval = TRUE---------------------------------------------
+## Generate data.
+set.seed(1986)
 
-## ----estimate-cates, eval = FALSE---------------------------------------------
-#  ## Sample split.
-#  splits <- sample_split(length(y), training_frac = 0.5)
-#  training_idx <- splits$training_idx
-#  honest_idx <- splits$honest_idx
-#  
-#  y_tr <- y[training_idx]
-#  D_tr <- D[training_idx]
-#  X_tr <- X[training_idx, ]
-#  
-#  y_hon <- y[honest_idx]
-#  D_hon <- D[honest_idx]
-#  X_hon <- X[honest_idx, ]
-#  
-#  ## Estimate the CATEs. Use only training sample.
-#  library(grf)
-#  forest <- causal_forest(X_tr, y_tr, D_tr)
-#  cates <- predict(forest, X)$predictions
+n <- 500 # Small sample size due to compliance with CRAN notes.
+k <- 3
 
-## ----construct-sequence, eval = FALSE-----------------------------------------
-#  ## Construct the sequence. Use doubly-robust scores.
-#  groupings <- build_aggtree(y, D, X, method = "aipw",
-#                             cates = cates, is_honest = 1:length(y) %in% honest_idx)
-#  
-#  ## Print.
-#  print(groupings)
-#  
-#  ## Plot.
-#  plot(groupings) # Try also setting 'sequence = TRUE'.
+X <- matrix(rnorm(n * k), ncol = k)
+colnames(X) <- paste0("x", seq_len(k))
+D <- rbinom(n, size = 1, prob = 0.5)
+mu0 <- 0.5 * X[, 1]
+mu1 <- 0.5 * X[, 1] + X[, 2]
+Y <- mu0 + D * (mu1 - mu0) + rnorm(n)
 
-## ----inference, eval = FALSE--------------------------------------------------
-#  ## Inference with 4 groups.
-#  results <- inference_aggtree(groupings, n_groups = 4)
-#  
-#  ## LATEX.
-#  print(results, table = "diff")
-#  print(results, table = "avg_char")
+## Sample split.
+splits <- sample_split(length(Y), training_frac = 0.5)
+training_idx <- splits$training_idx
+honest_idx <- splits$honest_idx
+
+Y_tr <- Y[training_idx]
+D_tr <- D[training_idx]
+X_tr <- X[training_idx, ]
+
+Y_hon <- Y[honest_idx]
+D_hon <- D[honest_idx]
+X_hon <- X[honest_idx, ]
+
+## ----estimate-cates, eval = TRUE----------------------------------------------
+## Estimate the CATEs. Use only training sample.
+forest <- causal_forest(X_tr, Y_tr, D_tr) 
+
+cates_tr <- predict(forest, X_tr)$predictions
+cates_hon <- predict(forest, X_hon)$predictions
+
+## ----construct-sequence, eval = TRUE------------------------------------------
+## Construct the sequence. Use doubly-robust scores (default option).
+groupings <- build_aggtree(Y_tr, D_tr, X_tr, # Training sample. 
+                           Y_hon, D_hon, X_hon, # Honest sample.
+                           cates_tr = cates_tr, cates_hon = cates_hon) # Predicted CATEs.
+
+## Print.
+print(groupings)
+
+## Plot.
+plot(groupings) # Try also setting 'sequence = TRUE'.
+
+## ----inference, eval = TRUE---------------------------------------------------
+## Inference with 4 groups.
+results <- inference_aggtree(groupings, n_groups = 4)
+
+## LATEX.
+print(results, table = "diff")
+
+print(results, table = "avg_char")