+++ NEWS.md
@@ -1,3 +1,11 @@
+## version 2.1.0
+
+The 'build_aggtree' function now requires a different set of arguments that allow better control on training-honest sample split.
+
+## version 2.0.3
+
+Updated vignettes. Improved LATEX output.
+
 ## version 2.0.2
 
 Fixed a bug that occurred when using only one covariate.